Refactoring the Login/Sign to Reusable API Calls and FrontEnd Properties
