+++ Front-End/package.json
@@ -19,12 +19,15 @@
     "react-circular-progressbar": "^2.1.0",
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
+    "react-draggable": "^4.4.5",
     "react-horizontal-scrolling-menu": "^3.2.5",
     "react-icons": "^4.7.1",
+    "react-lineto": "^3.3.0",
     "react-native-fs": "^2.20.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.0",
     "react-scripts": "^2.1.8",
+    "react-xarrows": "^2.0.2",
     "reactstrap": "^9.1.5",
     "recharts": "^2.4.3",
     "web-vitals": "^2.1.4"
