+++ Front-End/src/components/Functions/APIUrl.js
@@ -1,0 +1,2 @@
+export const signIn = "polyglot/player/signIn";
+export const signUp = "polyglot/player/signUp";
