+++ Front-End/src/components/SignupLogin/Signup.css
@@ -1,6 +1,3 @@
-/* First Divison Start */
-
-/* Main Division */
 .first-div {
   background-color: #002d40;
   height: 984px;
@@ -8,7 +5,6 @@
   padding-top: 2%;
 }
 
-/* h2 Tag */
 .h2-acc {
   color: white;
   font-weight: 900;
@@ -17,7 +13,6 @@
   margin-left: 15%;
 }
 
-/* p Tag */
 .p-acc {
   color: white;
   font-weight: 500;
@@ -29,23 +24,15 @@
   padding-top: 2%;
 }
 
-/* img Tag */
 .img-acc {
   width: 70%;
   padding-top: 5%;
 }
 
-/* First Divison End */
-
-/* Second Divison Start */
-
-/* Main Division */
 .second-div {
   background-color: rgba(26, 166, 183, 0.1);
   height: 984px;
 }
-
-/* Login/Signup Division */
 
 .sign-div {
   display: block;
@@ -60,26 +47,22 @@
   }
 }
 
-/* nav Tag border disable */
 .nav-tabs {
   border-bottom: none !important;
 }
 
-/* button Tag active */
 li button.active {
   background-color: #002d40 !important;
   color: white !important;
   border-radius: 10px !important;
 }
 
-/* button Tag hover */
 li button:hover {
   background-color: #002d40 !important;
   color: white !important;
   border-radius: 10px !important;
 }
 
-/* button Tag */
 li button {
   border: none !important;
   border-color: white !important;
@@ -89,7 +72,6 @@ li button {
   text-align: center;
 }
 
-/* title changes on hover */
 .title-beauty {
   font-weight: bolder;
   background-color: #002d40;
@@ -114,7 +96,6 @@ li button {
   text-align: center;
 }
 
-/* p Tag */
 .about-p {
   margin-top: 6%;
   font-size: 1.5rem;
@@ -123,7 +104,6 @@ li button {
   color: #1aa6b7;
 }
 
-/* input Tag */
 .input-acc {
   border-width: 0.1rem;
   border-radius: 30px;
@@ -168,14 +148,12 @@ li button {
   font-size: 1.5rem;
 }
 
-/* p Tag placement */
-.input-acc::placeholder {
+.input-acc:placeholder {
   font-size: 1rem;
   font-weight: 500;
   color: rgba(0, 45, 64, 0.6);
 }
 
-/* p Tag hover */
 .input-acc:hover {
   background-color: rgba(26, 166, 183, 0.3);
   border-color: #1aa6b7;
@@ -183,7 +161,6 @@ li button {
   font-weight: 500;
 }
 
-/* button Tag */
 .btn-acc {
   background-color: #002d40;
   padding: 1% 20%;
@@ -192,13 +169,11 @@ li button {
   border-radius: 30px;
 }
 
-/* button Division */
 .btn-div {
   margin-top: 4%;
   text-align: center;
 }
 
-/* division a Tag */
 .acc-div-a {
   text-align: center;
   padding-top: 5%;
@@ -206,7 +181,6 @@ li button {
   font-weight: 500;
 }
 
-/* a Tag */
 .a-acc {
   text-decoration: none;
   color: #1aa6b7;
@@ -214,12 +188,10 @@ li button {
   font-size: 1.1rem;
 }
 
-/* a Tag hover */
 .a-acc:hover {
   color: #002d40;
 }
 
-/* span Tag error */
 .span-acc-error {
   color: #ff414d;
   font-weight: 500;
@@ -227,7 +199,6 @@ li button {
   margin-left: 5%;
 }
 
-/* span Tag pass */
 .span-acc-pass {
   color: #1aa6b7;
   font-weight: 500;
@@ -235,7 +206,6 @@ li button {
   margin-left: 5%;
 }
 
-/* icon Tag */
 .icon-acc {
   margin-right: 2%;
 }
@@ -260,5 +230,3 @@ li button {
 div.alert {
   text-align: center;
 }
-
-/* Second Divison End */
